+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 ## v4.1.0
 - Initial release of PipeDream PowerShell module v4
 - Like Star Wars Episodes I, II and III we dont talk about v1, v2 and v3
+
+## v4.2.0
+- Added `Invoke-DataverseGet`
+- Added `Invoke-DataversePost`
+- Added `Invoke-DataversePatch`
+- Added `Invoke-DataverseDelete`